+++ packages/api-client/src/components/ApiClient/ApiClient.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { Tab, TabGroup, TabList, TabPanel, TabPanels } from '@headlessui/vue'
-import '@scalar/default-theme/scrollbar.css'
-import '@scalar/default-theme/theme.css'
+import '@scalar/themes/default.css'
+import '@scalar/themes/scrollbar.css'
 import { useKeyboardEvent } from '@scalar/use-keyboard-event'
 import { useMediaQuery } from '@vueuse/core'
 import { ref, watch } from 'vue'